backend/requirements.txt
gunicorn==19.0.0
django==2.2.5
djangorestframework==3.10.3
django-filter==2.2.0
markdown==3.3.3
coreapi==2.3.3
psycopg2==2.8.3
geoip2==0.2.0